GOVSI-1114: Extract common Dynamo extension code into abstract class
- Extract the common code between UserStoreExtension and ClientStoreExtension into a base abstract class
- Make UserStoreExtension and ClientStoreExtension extend the new base class
