+++ shared-test/src/main/java/uk/gov/di/authentication/sharedtest/extensions/ClientStoreExtension.java
@@ -1,43 +1,25 @@
 package uk.gov.di.authentication.sharedtest.extensions;
 
-import com.amazonaws.client.builder.AwsClientBuilder;
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
 import com.amazonaws.services.dynamodbv2.model.AttributeDefinition;
-import com.amazonaws.services.dynamodbv2.model.AttributeValue;
 import com.amazonaws.services.dynamodbv2.model.CreateTableRequest;
 import com.amazonaws.services.dynamodbv2.model.GlobalSecondaryIndex;
 import com.amazonaws.services.dynamodbv2.model.KeySchemaElement;
 import com.amazonaws.services.dynamodbv2.model.Projection;
-import com.amazonaws.services.dynamodbv2.model.ResourceNotFoundException;
-import com.amazonaws.services.dynamodbv2.model.ScanRequest;
-import com.amazonaws.services.dynamodbv2.model.ScanResult;
 import org.junit.jupiter.api.extension.AfterEachCallback;
-import org.junit.jupiter.api.extension.BeforeAllCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import uk.gov.di.authentication.shared.services.DynamoClientService;
-import uk.gov.di.authentication.shared.services.DynamoService;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 
 import static com.amazonaws.services.dynamodbv2.model.KeyType.HASH;
 import static com.amazonaws.services.dynamodbv2.model.ProjectionType.ALL;
 import static com.amazonaws.services.dynamodbv2.model.ScalarAttributeType.S;
 
-public class ClientStoreExtension implements BeforeAllCallback, AfterEachCallback {
-    private static final String REGION = System.getenv().getOrDefault("AWS_REGION", "eu-west-2");
-    private static final String ENVIRONMENT = System.getenv().getOrDefault("ENVIRONMENT", "local");
-    private static final String DYNAMO_ENDPOINT =
-            System.getenv().getOrDefault("DYNAMO_ENDPOINT", "http://localhost:8000");
-    private static final DynamoService DYNAMO_SERVICE =
-            new DynamoService(REGION, ENVIRONMENT, Optional.of(DYNAMO_ENDPOINT));
+public class ClientStoreExtension extends DynamoExtension implements AfterEachCallback {
 
     private final DynamoClientService dynamoClientService =
             new DynamoClientService(REGION, ENVIRONMENT, Optional.of(DYNAMO_ENDPOINT));
-
-    private AmazonDynamoDB dynamoDB;
 
     public void registerClient(
             String clientID,
@@ -67,32 +49,13 @@ public class ClientStoreExtension implements BeforeAllCallback, AfterEachCallbac
         return dynamoClientService.isValidClient(clientID);
     }
 
-    public void flushData() {
+    @Override
+    public void afterEach(ExtensionContext context) throws Exception {
         clearDynamoTable(dynamoDB, "local-client-registry", "ClientID");
     }
 
-    private void clearDynamoTable(AmazonDynamoDB dynamoDB, String tableName, String key) {
-        ScanRequest scanRequest = new ScanRequest().withTableName(tableName);
-        ScanResult result = dynamoDB.scan(scanRequest);
-
-        for (Map<String, AttributeValue> item : result.getItems()) {
-            dynamoDB.deleteItem(tableName, Map.of(key, item.get(key)));
-        }
-    }
-
     @Override
-    public void afterEach(ExtensionContext context) throws Exception {
-        flushData();
-    }
-
-    @Override
-    public void beforeAll(ExtensionContext context) throws Exception {
-        dynamoDB =
-                AmazonDynamoDBClientBuilder.standard()
-                        .withEndpointConfiguration(
-                                new AwsClientBuilder.EndpointConfiguration(DYNAMO_ENDPOINT, REGION))
-                        .build();
-
+    protected void createTables() {
         if (!tableExists("local-client-registry")) {
             createClientRegistryTable("local-client-registry");
         }
@@ -112,14 +75,5 @@ public class ClientStoreExtension implements BeforeAllCallback, AfterEachCallbac
                                         .withKeySchema(new KeySchemaElement("ClientName", HASH))
                                         .withProjection(new Projection().withProjectionType(ALL)));
         dynamoDB.createTable(request);
-    }
-
-    private boolean tableExists(String tableName) {
-        try {
-            dynamoDB.describeTable(tableName);
-            return true;
-        } catch (ResourceNotFoundException ignored) {
-            return false;
-        }
     }
 }
